More tests and miscs
+ Assign matrix to another matrix
+ Is matrix symetric
+ Is matrix upper triangular
+ Is matrix lower triangular
+ Make matrix upper triangular
+ Change all REQUIRE to CHECK so that it does not abort on failure
+ TODOs


See more on CHECK and REQUIRE here:
https://github.com/catchorg/Catch2/blob/devel/docs/assertions.md

The REQUIRE family of macros tests an expression and aborts the test case if it fails. The CHECK family are equivalent but execution continues in the same test case even if the assertion fails. This is useful if you have a series of essentially orthogonal assertions and it is useful to see all the results rather than stopping at the first failure.